Interfaces for guilds' services added, double fields changed to decimal

+++ AnkhMorporkApp/Guilds/GuildOfBeggars.cs
@@ -1,7 +1,5 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.IO;
-using System.Reflection;
 using Newtonsoft.Json;
 
 namespace AnkhMorporkApp
@@ -27,7 +25,7 @@ namespace AnkhMorporkApp
         public override void BalanceChange(Player player, Beggar beggar)
         {
             string number = null;
-            double input = 0;
+            decimal input = 0;
             Console.WriteLine($"You came across a beggar!\nEnter 's' to skip. Or give sum of {beggar.Fee}");
             if (player.IsOutOfMoney(beggar.Fee))
                 return;
@@ -40,12 +38,12 @@ namespace AnkhMorporkApp
                     player.Skip(number, beggar);
                     return;
                 }
-                if (!Double.TryParse(number, out double result))
+                if (!Decimal.TryParse(number, out decimal result))
                 {
                     Console.WriteLine("Incorrect data! Please, try again:");
                     continue;
                 }
-                input = Double.Parse(number);
+                input = Decimal.Parse(number);
                 if (input != beggar.Fee)
                     Console.WriteLine($"The amount isn't equal {beggar.Fee}! Please, try again:");
                 else if (!player.EnteredSumIsCorrect(input))
