+++ AnkhMorporkApp/Models/Assassin.cs
@@ -4,11 +4,11 @@ namespace AnkhMorporkApp
     public class Assassin
     {
         public string Name { get; set; }
-        public double MinReward { get; set; }
-        public double MaxReward { get; set; }
+        public decimal MinReward { get; set; }
+        public decimal MaxReward { get; set; }
         public bool IsOccupied { get; set; }
 
-        public Assassin(string name, double minReward, double maxReward, bool isOccupied)
+        public Assassin(string name, decimal minReward, decimal maxReward, bool isOccupied)
         {
             this.Name = name;
             this.MinReward = minReward;
