+++ AnkhMorporkApp/Models/Beggar.cs
@@ -4,9 +4,9 @@ namespace AnkhMorporkApp
     public class Beggar
     {
         public string Practice { get; set; }
-        public double Fee { get; set; }
+        public decimal Fee { get; set; }
 
-        public Beggar(string practice, double fee)
+        public Beggar(string practice, decimal fee)
         {
             this.Practice = practice;
             this.Fee = fee;
